requirements.txt
django
django-sass-processor
django-solo

gunicorn
dj_database_url
psycopg2
django_heroku
libsass
django-compressor
whitenoise